requirements.txt
requests
python-telegram-bot==13.7
openai
ffmpeg
pydub
python-dotenv
hupper
pillow